Move proxiable logic closer together for better locality of behavior
