--- packages/arbor-ost/src/visitors/$object/$proxiable.ts
@@ -1,12 +1,0 @@
-import { Visitor } from "../visitor"
-import { isProxiable } from "../../visitors"
-
-export class ProxiableVisitor extends Visitor {
-  accepts({ childValue }) {
-    return isProxiable(childValue)
-  }
-
-  visit({ ost, $node, childValue }) {
-    return ost.nodeOf(childValue) || $node.$createChild(childValue)
-  }
-}
